requirements.txt
streamlit==1.38.0
transformers==4.44.0
torch>=2.0.0
newspaper3k==0.2.8
lxml_html_clean==0.1.1